Move controllers in controller folder
