+++ src/main/java/com/blind/company/api/controller/CompanyController.java
@@ -1,4 +1,4 @@
-package com.blind.company.api;
+package com.blind.company.api.controller;
 
 import org.springframework.web.bind.annotation.RestController;
